Feature: Implemented Components with SCSS Styles for AI Movie Recommendation Page
In this commit, I created components for the page where artificial intelligence intelligently suggests movies. The logic for displaying and removing movies was developed, along with forms for creating and deleting collections and a form for adding movies to collections. Additionally, SCSS styles were implemented to enhance the visual presentation.